refactor(auth): rename app
