+++ aphorism/apps/__init__.py
@@ -2,6 +2,8 @@
 from flask import Blueprint
 from flask_restx import Api
 
+# FIXME: strange import order
+from aphorism.apps.auth.controllers import auth_ns
 from aphorism.apps.user.controllers import user_ns
 
 api_v1_blueprint = Blueprint("api_v1", __name__)
@@ -18,4 +20,5 @@ api_v1 = Api(
     authorizations=authorizations,
 )
 
+api_v1.add_namespace(auth_ns, path="/auth")
 api_v1.add_namespace(user_ns, path="/user")
